🔍 io_archive_seek takes unsigned long (absolute position)
The offset is always an absolute SEEK_SET position (window base + read cursor) —
a non-negative magnitude computed from and stored into unsigned long fields, not
a signed displacement like the CRT fseek. Type it unsigned long; the explicit
(long)offset cast marks where it converts for the signed fseek call. Doxygen
updated. Byte-identical for 1.00 and 1.02.

+++ bak/SRC/IO/IO.C
@@ -480,12 +480,12 @@ void bak_select_archive(int archiveIndex) {
     }
 }
 
-void io_archive_seek(long offset) {
+void io_archive_seek(unsigned long offset) {
     Archive *ar;
 
     ar = &g_ioArchives[g_ioCurrentArchive];
     if (ar->filePos != offset) {
-        fseek(ar->fp, offset, SEEK_SET);
+        fseek(ar->fp, (long)offset, SEEK_SET);
         ar->filePos = offset;
     }
 }
